requirements.txt
Flask==3.1.2
torch==2.10.0
torchvision==0.25.0
Pillow==12.1.0
numpy==2.4.1
